dspy/requirements.txt
dspy-ai
python-dotenv
openai
